Use random values for UIDVALIDITY instead of timestamps
Random values have much lesser chance of collision than second-percision
timestamps (we can't get more than that as we are limited by 32-bit
integer bounds). This allows us to do several things by... just, well,
not doing anything.

Say, we create treeof mailboxes INBOX, INBOX/a, INBOX/a/a1, INBOX/b2.
If we do thing within one second - this creates three mailboxes with the
same UIDs. Now, we drop INBOX/b2 and rename INBOX/a to INBOX/b2.
This requires us to reassign new UIDVALIDITY values for INBOX/a and
all children. Since there are still no standard syntax to get UNIX
timestamp in SQL... We have to do a pretty expensive iteration with a
lot of overhead.

But with random values we get all this for nothing - we just can't get
two mailboxes with the same UIDVALIDITY value. This simplifies things a
lot. We can just change mailbox name in database and that's all.