Create override imap folders if needed
The delivery allows for user overrides to affect the destination imap
folder.  The function documentation indicates that if the destination
folder does not exist, it will be created.  However, the creation
function only works for explicitly set folders, not for the overrides.

This change creates a simple `GetOrCreate` user call for an imap folder.
It then leverages this call both in the original non-override path as
well as the override path.